Make ScratchList loading less greedy
